@@ -15,63 +15,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/platform-server": "~8.2.14",
-    "@angular/router": "~8.2.14",
+    "@angular/animations": "~10.0.9",
+    "@angular/cdk": "^10.1.3",
+    "@angular/common": "~10.0.9",
+    "@angular/compiler": "~10.0.9",
+    "@angular/core": "~10.0.9",
+    "@angular/forms": "~10.0.9",
+    "@angular/platform-browser": "~10.0.9",
+    "@angular/platform-browser-dynamic": "~10.0.9",
+    "@angular/platform-server": "~10.0.9",
+    "@angular/router": "~10.0.9",
     "@angularclass/hmr": "^2.1.3",
-    "@briebug/cypress-schematic": "^3.0.2",
+    "@briebug/cypress-schematic": "^3.3.0",
     "@briebug/jest-schematic": "^2.1.1",
-    "@ngrx/entity": "^8.6.0",
-    "@ngrx/router-store": "^8.6.0",
-    "@ngrx/schematics": "^8.6.0",
-    "@ngrx/store": "^8.6.0",
-    "@ngrx/store-devtools": "^8.6.0",
+    "@ngrx/entity": "^10.0.0",
+    "@ngrx/router-store": "^10.0.0",
+    "@ngrx/schematics": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@ngrx/store-devtools": "^10.0.0",
     "@nguniversal/module-map-ngfactory-loader": "^8.2.6",
-    "ajv": "^6.10.2",
-    "babylonjs": "^4.0.3",
-    "core-js": "^2.5.4",
+    "ajv": "^6.12.4",
+    "babylonjs": "^4.1.0",
+    "core-js": "^3.6.5",
     "faker": "^4.1.0",
-    "js-yaml": "^3.13.1",
-    "marked": "^0.8.0",
-    "ngx-markdown": "^8.2.1",
-    "prismjs": "^1.17.1",
-    "rxjs": "~6.5.4",
+    "js-yaml": "^3.14.0",
+    "marked": "^1.1.1",
+    "ngx-markdown": "^10.1.0",
+    "prismjs": "^1.21.0",
+    "rxjs": "~6.6.2",
     "sass-breakpoints": "^2.1.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^8.3.2",
-    "@angular-devkit/build-angular": "~0.803.24",
-    "@angular/cli": "~8.3.24",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@cypress/webpack-preprocessor": "4.1.1",
-    "@types/faker": "^4.1.8",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/jest": "24.0.24",
-    "@types/node": "^12.12.21",
-    "codelyzer": "^5.0.1",
-    "cypress": "3.8.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "jest": "24.9.0",
-    "jest-marbles": "^2.5.0",
-    "jest-preset-angular": "^8.0.0",
-    "ng-mocks": "^8.1.0",
-    "ng-openapi-gen": "^0.7.0",
-    "ts-loader": "6.2.1",
-    "ts-node": "^8.5.4",
-    "tslib": "^1.10.0",
-    "tslint": "^5.20.1",
-    "typescript": "^3.5.3"
+    "@angular-builders/jest": "^10.0.0",
+    "@angular-devkit/build-angular": "~0.1000.6",
+    "@angular/cli": "~10.0.6",
+    "@angular/compiler-cli": "~10.0.9",
+    "@angular/language-service": "~10.0.9",
+    "@cypress/webpack-preprocessor": "5.4.4",
+    "@types/faker": "^4.1.12",
+    "@types/jasmine": "~3.5.12",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/jest": "26.0.10",
+    "@types/node": "^14.0.27",
+    "codelyzer": "^6.0.0",
+    "cypress": "4.12.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.2",
+    "jest": "26.4.0",
+    "jest-marbles": "^2.5.1",
+    "jest-preset-angular": "^8.2.1",
+    "ng-mocks": "^10.1.2",
+    "ng-openapi-gen": "^0.12.1",
+    "ts-loader": "8.0.2",
+    "ts-node": "^8.10.2",
+    "tslib": "^2.0.1",
+    "tslint": "^6.1.3",
+    "typescript": "^3.9.7"
   },
   "jest": {
     "preset": "jest-preset-angular",
