@@ -20,7 +20,7 @@ import { CommentsService } from './services/comments.service';
   ],
 })
 export class ApiModule {
-  static forRoot(params: ApiConfigurationParams): ModuleWithProviders {
+  static forRoot(params: ApiConfigurationParams): ModuleWithProviders<ApiModule> {
     return {
       ngModule: ApiModule,
       providers: [
